Remove checkfile endpoint and extract file checking functions to utilis.py
